Move typeOf to AbstractGitLabSCMHeadEvent
